Clarify Usage of `v-html`
Originally, I'd thought that:

```
<div v-html="rawHtml">This {{ some_html}}</div>
```

(where we might have `vm.data.some_html='<b>hello</b>'`) ... would render something like:

**hello**

But instead we get:

```
Property or method "rawHtml" is not defined on the instance but referenced during render. Make sure to declare reactive data properties in the data option.
```

...if one is watching console errors.

My wording could be a little bit kludgy and others might be able to make it clearer but I think this may overcome some misunderstandings for those not familiar with vuejs.